+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add util/atomic.h
 - `Logger` class to centralize CI runner script logging (in particular, indentation)
+- Explicit reporting of free bytes after compilation
 
 ### Changed
 - `arduino_ci.rb` uses new `Logger`